fix(ui): move the email tab helper out of the layout module
CI's eslint caught what my local checks did not: EmailLayout exported both
components and a plain function, which breaks fast refresh
(react-refresh/only-export-components). The helper now has its own module,
which is where a shared function belonged anyway.

I had been running prettier, tsc and vitest but never `pnpm lint` or
`pnpm build` — the two steps of the CI job that had anything left to say.
